requirements.txt
pyserial
flask
waitress